[BACKLOG-6547] - CDH55 secure shim: PUC: DB Table(s) Data Source cannot be created based on Secured Cloudera Impala JDBC connection
the issue is caused by the fact that two different sets of "extra" properties could be stored for a db connection in BI repository:
the first is under extraOptions node (BI style)
the second is in attributes collection with EXTRA_OPTION_ prefix (DI style)

The Database Dialog displays and changes the BI style options, while a real connection is established using the DI style one.
In order to fix this the DatabaseUtil.convertToDatabaseMeta() method is changed so that BI style options has higher precidence.
Checkstyle is fixed for the affected sources.
Unit Test is provided (ivy.xml has been chenged to resolve Mockito library)
License headers updated